+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Bug fixes
 ---------
 
 * API for registering property getters/setters and indexers to an `Engine` now works with functions that take a first parameter of `NativeCallContext`.
+* Missing API function `Module::set_getter_setter_fn` is added.
 
 
 Version 1.9.0